사용자 평점까지 마무리

+++ yesProject/src/main/java/com/bit/yes/model/ReviewDao.java
@@ -46,5 +46,5 @@ public interface ReviewDao {
 
 	int reviewEditComment(CommentVo commentVo);
 
-	int loadReviewScoreAvg(String branchId);
+	double loadReviewScoreAvg(String branchId);
 }
